Migration: Run the Create Invoices model migration
Invoice table is created in the database and matches the invoices model.
This branch and change is to capture the migration changes seperately to
all other code changes as per Hnry's way of working to ensure changes to
the database are in sync before we attempt to access it via the model.
This also ensures the changes can be rolled back much easier and is more
predictable for reviewers.