Model: Create Invoice Items model and associations
I created the InvoiceItems model and defined it's reference relationship
to the Invoice model from the Rails CLI such that an invoice has many
invoice items and invoice items belong to one or more invoices.

Rails automatically saves associated models when the parent model is
saved; i.e. invoice items are saved when their invoice is saved.
When invoice items are saved however, the associated invoice also needs
to be updated and then saved.

Added a method to the Invoice model to calculate and update the
invoice's new total when any of it's items are saved. This invoice model
method is called from the invoice items model using the `after_save`
callback which makes an explicit `invoice.save` call. This triggers the
`before_save`callback in `invoice`. The new total is calculated and
assigned via `self.total`, and there is a check for redundant total
updates. The explicit save that was called from invoice items continues,
and the recalculated invoice is saved.